# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Grief Navigator API
  version: 1.0.0

servers:
  - url: /api

paths:
  /loss/register:
    post:
      operationId: registerLosses
      summary: Register losses
      tags:
        - Loss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/LossDTO"
      responses:
        "200":
          description: Loss registered successfully
        "400":
          description: Bad request

  /loss:
    get:
      operationId: getAllLosses
      summary: Get all losses for a perticular user
      tags:
        - Loss
      responses:
        "200":
          description: A list of losses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LossDTO"
        "500":
          description: Internal server error

  /auth/me:
    get:
      operationId: getMyAppUser
      summary: Register a new app user
      tags:
        - AppUser
      responses:
        "200":
          description: App user registered successfully
        "400":
          description: Bad request

  /difficultTimes:
    post:
      operationId: registerDifficultTimes
      summary: Save difficult difficultTimes
      tags:
        - DifficultTimes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DifficultTimes"
      responses:
        "200":
          description: Difficulties acknowleged
        "400":
          description: Bad request
  
  /beliefStatement:
    get:
      operationId: getBeliefStatements
      summary: Fetch pre defined beliefStatement
      tags:
        - BeliefStatement
      responses:
        "200":
          description: Fetch beliefStatement
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: string
        "400":
          description: Bad request
              
    post:
      operationId: saveBeliefStatements
      summary: Save beliefStatement
      tags:
        - BeliefStatement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: Fetch beliefStatement
        "400":
          description: Bad request
  
  /beliefEntry:
    post:
      operationId: saveBeliefEntries
      summary: Save Belief Entries for a user
      tags:
        - BeliefEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: 
                $ref: '#/components/schemas/BeliefEntryDTO'
      responses:
        '200':
          description: Entires Saved successfully
        '400':
          description: Bad Request

  /avoidenceStatement:
    post:
      operationId: saveAvoidenceStatements
      tags:
        - Avoidence Statement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: "Avoidence Statements saved successfully"
        "400":
          description: "Bad request"

    get:
      operationId: getAvoidenceStatements
      tags:
        - Avoidence Statement
      responses:
        "200":
          description: "Avoidence Statement fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: "Bad Request"

  /avoidenceEntry:
    post:
      operationId: saveAvoidenceEntries
      tags:
        - Avoidence Entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AvoidenceEntryDTO'
      responses:
        "200":
          description: "Avoidence Entires saved successfully"
        "400":
          description: "Bad Request"
  
  /familyConflict/{sectionId}:
    get:
      operationId: getQuestions
      tags:
        - Family Conflict 
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
        - name: fieldId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "Family Conflict Questions fetched successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyConflictQuestionDTO'
        "400":
          description: "Bad Request"
  
  /familyConflict:
    post:
      operationId: saveQuestions
      tags:
        - Family Conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/FamilyConflictQuestionDTO'
      responses:
        "200":
          description: "Saved Questions successfully"
        "400": 
          description: "Bad Request"
  
  /familyConflictAssessment/start:
    post:
      operationId: startAssessment
      tags:
        - Family Conflict Assessment 
      responses:
        "200":
          description: "Assessment started successfully"
          content:
            application/json:
              schema:
                type: long
        "400":
          description: "Bad Request"

  /familyConflictAssessment:
    get:
      operationId: getAssessment
      tags:
        - Family Conflict Assessment
      responses:
        "200":
          description: "Assessment fetched successfully"
          content:
            application/json:
              schema:
                type: object
        "400":
          description: "Bad Request"
  
  /familyConflictAssessment/complete:
    patch:
      operationId: completeAssessment
      tags:
        - Family Conflict Assessment
      parameters:
        - name: assessmentId
          in: query
          schema:
            type: long
      responses:
        "200":
          description: "Assessment Completed successfully"
        "400":
          description: "Bad Request"

components:
  schemas:
    FamilyConflictQuestionDTO:
      type: object
      properties:
        sectionId:
          type: string
        fieldId:
          type: string
        question:
          type: string
        type:
          $ref: '#/components/schemas/QuestionType'
        options:
          type: array
          items:
            type: string
      required:
        - fieldId
        - question
        - type
        - sectionId
    AvoidenceEntryDTO:
      type: object
      properties:
        avoidenceStatementId:
          type: long
        rating:
          type: integer
        relatedLosses:
          type: array
          items:
            type: long
    BeliefEntryDTO:
      type: object
      properties:
        beliefStatementId: 
          type: long
        rating:
          type: integer
        relatedLosses:
          type: array
          items: 
            type: long
    DifficultTimes:
      type: object
      properties:
        dayOrTime:
          type: string
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        related_loss:
          type: long
    LossDTO:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/LossType"
        description:
          type: string
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        time:
          type: string
      required:
        - type
        - description
        - difficulty
        - time
    QuestionType:
      type: string
      enum:
        - DESCRIPTIVE
        - RADIO
        - MCQ
        - MCQ_MULTI
    LossType:
      type: string
      enum:
        - RELATIONSHIP
        - THING
        - IDENTITY
    DifficultyLevel:
      type: string
      enum:
        - ZERO
        - ONE
        - TWO
        - THREE
        - FOUR